Backend codebase refactoring tro make it more modular and easy to work on for future tasks
