+++ models/schemas.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class ChatRequest(BaseModel):
+    user_query: str
+
+
+
