+++ modules/classification/classifier.py
@@ -10,7 +10,7 @@
 # model = AutoModelForCausalLM.from_pretrained(MODEL_NAME).to(device)
 
 from openai import OpenAI
-from config import OPENAI_API_KEY
+from core.config import OPENAI_API_KEY
 
 client = OpenAI(api_key=OPENAI_API_KEY)
